Split parser tests into multiple files
